refactor(sql insights): add the pure SQL pie data layer
Stand up sqlPieGraphAdapter.ts as the SQL pie's pure (non-React) data layer,
mirroring sqlLineGraphAdapter:
- buildPieSlices / formatPieSliceCount move out of PieChart.tsx (the chart.js
  wrapper imports them) — no behavior change
- buildPieSeries (slice -> quill Series) and the canRenderSqlPieGraph gate are
  added for the upcoming quill renderer

Each function is parameterized-tested; nothing is wired into a component yet, so
the rendered pie is unchanged. The quill SqlPieGraph + wrapper gating land on top.

Generated-By: PostHog Code
Task-Id: a725d77c-5810-42a7-be0a-a5e6bacc532b
